% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_assoc.R
\name{mixed_assoc}
\alias{mixed_assoc}
\title{Calculate correlations between different types of variables}
\usage{
mixed_assoc(df, cor_method = "pearson", adjust_cramersv_bias = TRUE)
}
\arguments{
\item{df}{a two-columns data frame (x and y respectively).}

\item{cor_method}{method to use for correlation between two quantitative variables}

\item{adjust_cramersv_bias}{do bias adjustment?}
}
\value{
This function returns a 2-elements list with:
\itemize{
\item a numeric giving the value on x-Axis corresponding to the optimal
\item a data frame with the original values and two additional columns used in
the graphic.
}
}
\description{
Calculate correlations between different types of variables
}
